Renaming neteq_opus_fec_quality_test.
neteq_opus_fec_quality_test has been modified to test more configurations of Opus than only FEC. It makes sense to rename it to neteq_opus_quality_test. This was planned in

https://webrtc-codereview.appspot.com/45619004/

but was forgotten. This CL handles it, and makes it easy for review.

BUG=
R=henrik.lundin@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/45709004